# pay-skills/providers/khun/thai-services/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Khun",
    "version": "0.1.0",
    "description": "Khun is the Thai supply side of Pay.sh. Thai service providers (tuk-tuk drivers, market vendors, tour guides, freelancers) onboard themselves via a public Telegram bot and become Solana mainnet AI agents on the agent-registry-8004. Foreign AI assistants discover them via Khun search, pay in Tether USDT via x402, and route the order back to the merchant on Telegram.",
    "contact": {
      "name": "Khun",
      "url": "https://github.com/ss251/khun"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://aquatic-entertainment-simulation-brunette.trycloudflare.com"
    }
  ],
  "paths": {
...
    "/agent/{agentId}": {
      "get": {
        "summary": "Read full agent record",
        "operationId": "getAgent",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "KhunAgent JSON"
          },
          "404": {
            "description": "Unknown agent"
          }
        }
      }
    },
...
  }
}